requirements.txt
azure-identity==1.6.1
azure-keyvault-secrets==4.3.0